[rbp/omxplayer] Check total memory, and if plenty use full size buffers
Part 2: Inplementation

omxplayer has used a quarter sized video buffer and a half sized audio buffer for the file reading message queue, compared to dvdplayer.
This is required on 256M boards, but is unneccessary on 512M board. So we query the memory size and use full size buffers if possible.