[rbp/omxplayer] Avoid submitting too large audio buffers
The audio_decode output port buffers are 32K and we want to ensure our packets fit in these.
Audio decode produces samples of 16 bits and rounds the channels up to the next power of two.
We need to calculate the largest input packet size for the given bpp and channels and limit submitted buffers to this.

This has been observed to cause no audio output with wma audio.